+++ cli/aws_orbit/remote_files/kubectl.py
@@ -16,7 +16,7 @@ import logging
 import os
 import shutil
 import time
-from typing import Any, Dict, List, Optional, Tuple
+from typing import Any, Dict, List, Optional, Tuple, cast
 
 import aws_orbit
 from aws_orbit import ORBIT_CLI_ROOT, exceptions, k8s, sh, utils
@@ -239,6 +239,7 @@ def _generate_kube_system_manifest(context: "Context", clean_up: bool = True) ->
             content,
             dict(
                 account_id=context.account_id,
+                region=context.region,
                 env_name=context.name,
                 cluster_name=f"orbit-{context.name}",
                 sts_ep="legacy" if context.networking.data.internet_accessible else "regional",
@@ -551,8 +552,27 @@ def deploy_env(context: "Context") -> None:
 
         # Patch Kubeflow
         _logger.debug("Orbit applying KubeFlow patch")
-        sh.run(f'kubectl patch deployment jupyter-web-app-deployment --patch "{patch}" -n kubeflow')
+        sh.run(f'kubectl patch deployment -n kubeflow jupyter-web-app-deployment --patch "{patch}"')
         sh.run("kubectl rollout restart deployment jupyter-web-app-deployment -n kubeflow")
+
+        # Patch Pods to push into Fargate when deploying in an isolated subnet
+        if not context.networking.data.internet_accessible:
+            patch = '{"spec":{"template":{"metadata":{"labels":{"orbit/node-type":"fargate"}}}}}'
+            sh.run(f"kubectl patch deployment -n istio-system authzadaptor --patch '{patch}'")
+
+            patch = (
+                '{"spec":{"template":{"metadata":{"labels":{"orbit/node-type":"fargate"}},'
+                '"spec":{"containers":[{"name":"alb-ingress-controller","args":["--ingress-class=alb"'
+                ',"--cluster-name=$(CLUSTER_NAME)","--aws-vpc-id=VPC_ID"]}]}}}}'
+            )
+            patch = patch.replace("VPC_ID", cast(str, context.networking.vpc_id))
+            sh.run(f"kubectl patch deployment -n kubeflow alb-ingress-controller --patch '{patch}'")
+
+            patch = (
+                '[{"op": "add", "path": "/spec/template/metadata/labels/orbit~1node-type", "value": "fargate"}, '
+                '{"op": "replace", "path": "/spec/template/spec/nodeSelector", "value": {}}]'
+            )
+            sh.run(f"kubectl patch deployment -n orbit-system landing-page-service --type json --patch '{patch}'")
 
         # Confirm env Service Endpoints
         confirm_endpoints(name="landing-page-service", namespace="orbit-system")