+++ cli/aws_orbit/remote_files/utils.py
@@ -13,7 +13,7 @@ def get_k8s_context(context: "Context") -> str:
     try:
         contexts: List[str] = [str(c["name"]) for c in config.list_kube_config_contexts()[0]]
     except config.config_exception.ConfigException as e:
-        _logger.debug("Context not found")
+        _logger.exception("Context not found")
         raise e
     expected_domain: str = f"@orbit-{context.name}.{context.region}.eksctl.io"
     for ctx in contexts: